An S4 class used to represent LC-MS datasets

MetaboSet is the main class used to represent data in the notame package. It is built upon the ExpressionSet class from the Biobase package. In addition to the slots inherited from ExpressionSet, MetaboSet has three slots of its own. The extra three slots hold special column names that are stored purely for convenience, as many functions use these as defaults.

Usage

# S4 method for class 'MetaboSet'
combined_data(object)

# S4 method for class 'MetaboSet'
group_col(object)

# S4 method for class 'MetaboSet'
group_col(object) <- value

# S4 method for class 'MetaboSet'
time_col(object)

# S4 method for class 'MetaboSet'
time_col(object) <- value

# S4 method for class 'MetaboSet'
subject_col(object)

# S4 method for class 'MetaboSet'
subject_col(object) <- value

# S4 method for class 'MetaboSet'
flag(object)

# S4 method for class 'MetaboSet'
flag(object) <- value

# S4 method for class 'MetaboSet,data.frame'
join_fData(object, dframe)

# S4 method for class 'MetaboSet,data.frame'
join_pData(object, dframe)

Arguments

object

A MetaboSet object.

value

string, name of column to be designated for holding group labels

dframe

a data frame with the new information

Methods (by generic)

combined_data(MetaboSet): Retrieve both sample information and features

group_col(MetaboSet): get name of column for group labels

group_col(MetaboSet) <- value: set name of column for group labels

time_col(MetaboSet): get name of column for time points

time_col(MetaboSet) <- value: set name of column for time points

subject_col(MetaboSet): get name of column for subject subject identifiers

subject_col(MetaboSet) <- value: set name of column for subject identifiers

flag(MetaboSet): get flags

flag(MetaboSet) <- value: set flags

join_fData(object = MetaboSet, dframe = data.frame): join new information to feature data

join_pData(object = MetaboSet, dframe = data.frame): join new information to pheno data

Slots

group_col

character, name of the column holding group information

time_col

character, name of the column holding time points

subject_col

character, name of the column holding subject identifiers

Constructor

See construct_metabosets for constructor function.