Show citations

This function lists citations behind the notame functions that have been called during the session. All notame functions update the list automatically. The citations are taken from the call to 'citation("package"), and complemented with a brief description of what the package was used for. NOTE: the citations might not point to the correct paper if the package authors have not supplied correct citation information for their package. The output is written to the current log file, if specified.

Usage

citations()

Value

None, the function is invoked for its side effect.

Examples

citations()
#> INFO [2025-10-24 13:52:32] Preprocessing and analyses were performed using notame package:
#> INFO [2025-10-24 13:52:32] To cite package ‘notame’ in publications use:
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   Klavus, A.; Kokla, M.; Noerman, S.; Koistinen, V.M.; Tuomainen, M.;
#> INFO [2025-10-24 13:52:32]   Zarei, I.; Meuronen, T.; Hakkinen, M.R.; Rummukainen, S.; Farizah
#> INFO [2025-10-24 13:52:32]   Babu, A.; Sallinen, T.; Karkkainen, O.; Paananen, J.; Broadhurst, D.;
#> INFO [2025-10-24 13:52:32]   Brunius, C.; Hanhineva, K. "notame": Workflow for Non-Targeted LC-MS
#> INFO [2025-10-24 13:52:32]   Metabolic Profiling. Metabolites 2020, 10, 135.
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   @Article{,
#> INFO [2025-10-24 13:52:32]     title = {"notame": Workflow for Non-Targeted LC-MS Metabolic Profiling},
#> INFO [2025-10-24 13:52:32]     author = {Anton Klavus and Marietta Kokla and Stefania Noerman and Ville M. Koistinen and Marjo Tuomainen and Iman Zarei and Topi Meuronen and Merja R. Hakkinen and Soile Rummukainen and Ambrin {Farizah Babu} and Taisa Sallinen and Olli Karkkainen and Jussi Paananen and David Broadhurst and Carl Brunius and Kati Hanhineva},
#> INFO [2025-10-24 13:52:32]     year = {2020},
#> INFO [2025-10-24 13:52:32]     journal = {Metabolites},
#> INFO [2025-10-24 13:52:32]     volume = {10},
#> INFO [2025-10-24 13:52:32]     number = {135},
#> INFO [2025-10-24 13:52:32]     url = {https://doi.org/10.3390/metabo10040135},
#> INFO [2025-10-24 13:52:32]     doi = {10.3390/metabo10040135},
#> INFO [2025-10-24 13:52:32]   }
#> INFO [2025-10-24 13:52:32] The primary data structure in notame is SummarizedExperiment:
#> INFO [2025-10-24 13:52:32] To cite package ‘SummarizedExperiment’ in publications use:
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   Morgan M, Obenchain V, Hester J, Pagès H (2025).
#> INFO [2025-10-24 13:52:32]   _SummarizedExperiment: A container (S4 class) for matrix-like
#> INFO [2025-10-24 13:52:32]   assays_. doi:10.18129/B9.bioc.SummarizedExperiment
#> INFO [2025-10-24 13:52:32]   <https://doi.org/10.18129/B9.bioc.SummarizedExperiment>, R package
#> INFO [2025-10-24 13:52:32]   version 1.39.2,
#> INFO [2025-10-24 13:52:32]   <https://bioconductor.org/packages/SummarizedExperiment>.
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   @Manual{,
#> INFO [2025-10-24 13:52:32]     title = {SummarizedExperiment: A container (S4 class) for matrix-like assays},
#> INFO [2025-10-24 13:52:32]     author = {Martin Morgan and Valerie Obenchain and Jim Hester and Hervé Pagès},
#> INFO [2025-10-24 13:52:32]     year = {2025},
#> INFO [2025-10-24 13:52:32]     note = {R package version 1.39.2},
#> INFO [2025-10-24 13:52:32]     url = {https://bioconductor.org/packages/SummarizedExperiment},
#> INFO [2025-10-24 13:52:32]     doi = {10.18129/B9.bioc.SummarizedExperiment},
#> INFO [2025-10-24 13:52:32]   }
#> INFO [2025-10-24 13:52:32] visualizations in notame are built with ggplot2:
#> INFO [2025-10-24 13:52:32] To cite ggplot2 in publications, please use
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
#> INFO [2025-10-24 13:52:32]   Springer-Verlag New York, 2016.
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:32] 
#> INFO [2025-10-24 13:52:32]   @Book{,
#> INFO [2025-10-24 13:52:32]     author = {Hadley Wickham},
#> INFO [2025-10-24 13:52:32]     title = {ggplot2: Elegant Graphics for Data Analysis},
#> INFO [2025-10-24 13:52:32]     publisher = {Springer-Verlag New York},
#> INFO [2025-10-24 13:52:32]     year = {2016},
#> INFO [2025-10-24 13:52:32]     isbn = {978-3-319-24277-4},
#> INFO [2025-10-24 13:52:32]     url = {https://ggplot2.tidyverse.org},
#> INFO [2025-10-24 13:52:32]   }
data(toy_notame_set)
ex_set <- flag_quality(toy_notame_set)
#> INFO [2025-10-24 13:52:33] 
#> 92% of features flagged for low quality
# Broadhurst et al.(2018) added to citations
citations()
#> INFO [2025-10-24 13:52:33] Preprocessing and analyses were performed using notame package:
#> INFO [2025-10-24 13:52:33] To cite package ‘notame’ in publications use:
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   Klavus, A.; Kokla, M.; Noerman, S.; Koistinen, V.M.; Tuomainen, M.;
#> INFO [2025-10-24 13:52:33]   Zarei, I.; Meuronen, T.; Hakkinen, M.R.; Rummukainen, S.; Farizah
#> INFO [2025-10-24 13:52:33]   Babu, A.; Sallinen, T.; Karkkainen, O.; Paananen, J.; Broadhurst, D.;
#> INFO [2025-10-24 13:52:33]   Brunius, C.; Hanhineva, K. "notame": Workflow for Non-Targeted LC-MS
#> INFO [2025-10-24 13:52:33]   Metabolic Profiling. Metabolites 2020, 10, 135.
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   @Article{,
#> INFO [2025-10-24 13:52:33]     title = {"notame": Workflow for Non-Targeted LC-MS Metabolic Profiling},
#> INFO [2025-10-24 13:52:33]     author = {Anton Klavus and Marietta Kokla and Stefania Noerman and Ville M. Koistinen and Marjo Tuomainen and Iman Zarei and Topi Meuronen and Merja R. Hakkinen and Soile Rummukainen and Ambrin {Farizah Babu} and Taisa Sallinen and Olli Karkkainen and Jussi Paananen and David Broadhurst and Carl Brunius and Kati Hanhineva},
#> INFO [2025-10-24 13:52:33]     year = {2020},
#> INFO [2025-10-24 13:52:33]     journal = {Metabolites},
#> INFO [2025-10-24 13:52:33]     volume = {10},
#> INFO [2025-10-24 13:52:33]     number = {135},
#> INFO [2025-10-24 13:52:33]     url = {https://doi.org/10.3390/metabo10040135},
#> INFO [2025-10-24 13:52:33]     doi = {10.3390/metabo10040135},
#> INFO [2025-10-24 13:52:33]   }
#> INFO [2025-10-24 13:52:33] The primary data structure in notame is SummarizedExperiment:
#> INFO [2025-10-24 13:52:33] To cite package ‘SummarizedExperiment’ in publications use:
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   Morgan M, Obenchain V, Hester J, Pagès H (2025).
#> INFO [2025-10-24 13:52:33]   _SummarizedExperiment: A container (S4 class) for matrix-like
#> INFO [2025-10-24 13:52:33]   assays_. doi:10.18129/B9.bioc.SummarizedExperiment
#> INFO [2025-10-24 13:52:33]   <https://doi.org/10.18129/B9.bioc.SummarizedExperiment>, R package
#> INFO [2025-10-24 13:52:33]   version 1.39.2,
#> INFO [2025-10-24 13:52:33]   <https://bioconductor.org/packages/SummarizedExperiment>.
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   @Manual{,
#> INFO [2025-10-24 13:52:33]     title = {SummarizedExperiment: A container (S4 class) for matrix-like assays},
#> INFO [2025-10-24 13:52:33]     author = {Martin Morgan and Valerie Obenchain and Jim Hester and Hervé Pagès},
#> INFO [2025-10-24 13:52:33]     year = {2025},
#> INFO [2025-10-24 13:52:33]     note = {R package version 1.39.2},
#> INFO [2025-10-24 13:52:33]     url = {https://bioconductor.org/packages/SummarizedExperiment},
#> INFO [2025-10-24 13:52:33]     doi = {10.18129/B9.bioc.SummarizedExperiment},
#> INFO [2025-10-24 13:52:33]   }
#> INFO [2025-10-24 13:52:33] visualizations in notame are built with ggplot2:
#> INFO [2025-10-24 13:52:33] To cite ggplot2 in publications, please use
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
#> INFO [2025-10-24 13:52:33]   Springer-Verlag New York, 2016.
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33] A BibTeX entry for LaTeX users is
#> INFO [2025-10-24 13:52:33] 
#> INFO [2025-10-24 13:52:33]   @Book{,
#> INFO [2025-10-24 13:52:33]     author = {Hadley Wickham},
#> INFO [2025-10-24 13:52:33]     title = {ggplot2: Elegant Graphics for Data Analysis},
#> INFO [2025-10-24 13:52:33]     publisher = {Springer-Verlag New York},
#> INFO [2025-10-24 13:52:33]     year = {2016},
#> INFO [2025-10-24 13:52:33]     isbn = {978-3-319-24277-4},
#> INFO [2025-10-24 13:52:33]     url = {https://ggplot2.tidyverse.org},
#> INFO [2025-10-24 13:52:33]   }
#> INFO [2025-10-24 13:52:33] Quality metrics were computed as per guidelines in:
#> INFO [2025-10-24 13:52:33] [1] "Broadhurst, David et al. Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies. Metabolomics : Official journal of the Metabolomic Society vol. 14,6 (2018): 72. doi:10.1007/s11306-018-1367-3"