Cohen's D

Computes Cohen's D for each feature. If time and ID are supplied, change between two time points is computed for each subject, and Cohen's d is computed from the changes.

Usage

cohens_d(object, group, id = NULL, time = NULL, assay.type = NULL)

Arguments

object

a SummarizedExperiment or MetaboSet object

group

character, name of the group column

id

character, name of the subject ID column

time

character, name of the time column

assay.type

character, assay to be used in case of multiple assays

Value

A data frame with Cohen's d for each feature.

Examples

data(example_set)
d_results <- cohens_d(drop_qcs(example_set), group = "Group")
#> INFO [2025-06-23 22:36:28] Starting to compute Cohen's D between groups B & A
#> INFO [2025-06-23 22:36:28] Cohen's D computed.
d_results_time <- cohens_d(drop_qcs(example_set),
  group = "Group", time = "Time", id = "Subject_ID"
)
#> INFO [2025-06-23 22:36:28] Starting to compute Cohen's D between groups B & A from time change 2 - 1
#> INFO [2025-06-23 22:36:28] Cohen's D computed.