Save plot to chosen format

Saves the given plot to a file. Supports pdf, svg, emf, png and tiff formats. If an error occurs with the plot, an empty file is created.

Usage

save_plot(p, file, ...)

Arguments

p

a ggplot object

file

the file path

...

other arguments to plot function, like width and height

Value

None, the function is invoked for its plot-saving side effect.

See also

Examples

data(example_set)
lm_results <- perform_lm(drop_qcs(example_set), 
  formula_char = "Feature ~ Group")
#> INFO [2025-06-23 22:38:27] Starting linear regression.
#> INFO [2025-06-23 22:38:27] Linear regression performed.

p <- volcano_plot(lm_results,
  x = "GroupB_Estimate",
  p = "GroupB_P", p_fdr = "GroupB_P_FDR",
  label = "Feature_ID",
  fdr_limit = 0.1
)
#> Warning: None of the FDR-adjusted p-values are below the significance level, not plotting the horizontal line.

save_plot(p, file = "test.pdf")
#> INFO [2025-06-23 22:38:28] Saved to: test.pdf