Statistics cleaning

Uses regexp to remove unnecessary columns from statistics results data frame. Can also rename columns effectively.

Usage

summarize_results(
  df,
  remove = c("Intercept", "CI95", "Std_error", "t_value", "z_value", "R2"),
  rename = NULL,
  summary = TRUE,
  p_limit = 0.05,
  fdr = TRUE
)

Arguments

df

data frame, statistics results

remove

list, should contain strings that are matching to unwanted columns

rename

named list, names should contain matches that are replaced with values

summary

logical, should summary columns be added

p_limit

numeric, limit for p-values to be counted

fdr

logical, should summary be done with fdr-fixed values

Value

A data frame with removed and/or renamed columns.

Examples

data(example_set)
# Simple manipulation to linear model results
lm_results <- perform_lm(drop_qcs(example_set), 
  formula_char = "Feature ~ Group + Time")
#> INFO [2025-06-23 22:38:31] Starting linear regression.
#> INFO [2025-06-23 22:38:32] Linear regression performed.
lm_results <-summarize_results(lm_results,
  rename = c("GroupB" = "GroupB_vs_A", "Time2" = "Time2_vs_1")
)