Finish a log

Logs the current date and time and session info, and switches logging off.

Usage

finish_log()

Value

None, the function is invoked for its side effect.

See also

Examples

finish_log()
#> INFO [2025-06-23 22:36:32] Finished analysis. Mon Jun 23 22:36:32 2025
#> Session info:
#> 
#> INFO [2025-06-23 22:36:32] R version 4.5.1 (2025-06-13)
#> INFO [2025-06-23 22:36:32] Platform: x86_64-pc-linux-gnu
#> INFO [2025-06-23 22:36:32] Running under: Ubuntu 24.04.2 LTS
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] Matrix products: default
#> INFO [2025-06-23 22:36:32] BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
#> INFO [2025-06-23 22:36:32] LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.26.so;  LAPACK version 3.12.0
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] locale:
#> INFO [2025-06-23 22:36:32]  [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8       
#> INFO [2025-06-23 22:36:32]  [4] LC_COLLATE=C           LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8   
#> INFO [2025-06-23 22:36:32]  [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C          
#> INFO [2025-06-23 22:36:32] [10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C   
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] time zone: UTC
#> INFO [2025-06-23 22:36:32] tzcode source: system (glibc)
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] attached base packages:
#> INFO [2025-06-23 22:36:32] [1] stats4    stats     graphics  grDevices utils     datasets  methods  
#> INFO [2025-06-23 22:36:32] [8] base     
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] other attached packages:
#> INFO [2025-06-23 22:36:32]  [1] notame_0.99.2               SummarizedExperiment_1.38.1
#> INFO [2025-06-23 22:36:32]  [3] GenomicRanges_1.60.0        GenomeInfoDb_1.44.0        
#> INFO [2025-06-23 22:36:32]  [5] IRanges_2.42.0              S4Vectors_0.46.0           
#> INFO [2025-06-23 22:36:32]  [7] MatrixGenerics_1.20.0       matrixStats_1.5.0          
#> INFO [2025-06-23 22:36:32]  [9] magrittr_2.0.3              ggplot2_3.5.2              
#> INFO [2025-06-23 22:36:32] [11] Biobase_2.68.0              BiocGenerics_0.54.0        
#> INFO [2025-06-23 22:36:32] [13] generics_0.1.4             
#> INFO [2025-06-23 22:36:32] 
#> INFO [2025-06-23 22:36:32] loaded via a namespace (and not attached):
#> INFO [2025-06-23 22:36:32]  [1] tidyselect_1.2.1        viridisLite_0.4.2       dplyr_1.1.4            
#> INFO [2025-06-23 22:36:32]  [4] farver_2.1.2            fastmap_1.2.0           digest_0.6.37          
#> INFO [2025-06-23 22:36:32]  [7] lifecycle_1.0.4         compiler_4.5.1          rlang_1.1.6            
#> INFO [2025-06-23 22:36:32] [10] sass_0.4.10             tools_4.5.1             igraph_2.1.4           
#> INFO [2025-06-23 22:36:32] [13] yaml_2.3.10             knitr_1.50              lambda.r_1.2.4         
#> INFO [2025-06-23 22:36:32] [16] askpass_1.2.1           S4Arrays_1.8.1          labeling_0.4.3         
#> INFO [2025-06-23 22:36:32] [19] htmlwidgets_1.6.4       curl_6.4.0              DelayedArray_0.34.1    
#> INFO [2025-06-23 22:36:32] [22] xml2_1.3.8              RColorBrewer_1.1-3      abind_1.4-8            
#> INFO [2025-06-23 22:36:32] [25] BiocParallel_1.42.1     Rtsne_0.17              withr_3.0.2            
#> INFO [2025-06-23 22:36:32] [28] purrr_1.0.4             desc_1.4.3              grid_4.5.1             
#> INFO [2025-06-23 22:36:32] [31] fansi_1.0.6             scales_1.4.0            cli_3.6.5              
#> INFO [2025-06-23 22:36:32] [34] rmarkdown_2.29          crayon_1.5.3            ragg_1.4.0             
#> INFO [2025-06-23 22:36:32] [37] httr_1.4.7              cachem_1.1.0            parallel_4.5.1         
#> INFO [2025-06-23 22:36:32] [40] BiocManager_1.30.26     formatR_1.14            XVector_0.48.0         
#> INFO [2025-06-23 22:36:32] [43] vctrs_0.6.5             Matrix_1.7-3            jsonlite_2.0.0         
#> INFO [2025-06-23 22:36:32] [46] ggrepel_0.9.6           systemfonts_1.2.3       fontawesome_0.5.3      
#> INFO [2025-06-23 22:36:32] [49] jquerylib_0.1.4         glue_1.8.0              pkgdown_2.1.3          
#> INFO [2025-06-23 22:36:32] [52] codetools_0.2-20        stringi_1.8.7           gtable_0.3.6           
#> INFO [2025-06-23 22:36:32] [55] futile.logger_1.4.3     UCSC.utils_1.4.0        downlit_0.4.4          
#> INFO [2025-06-23 22:36:32] [58] tibble_3.3.0            pillar_1.10.2           rappdirs_0.3.3         
#> INFO [2025-06-23 22:36:32] [61] pcaMethods_2.0.0        htmltools_0.5.8.1       openssl_2.3.3          
#> INFO [2025-06-23 22:36:32] [64] GenomeInfoDbData_1.2.14 R6_2.6.1                httr2_1.1.2            
#> INFO [2025-06-23 22:36:32] [67] textshaping_1.0.1       evaluate_1.0.4          lattice_0.22-7         
#> INFO [2025-06-23 22:36:32] [70] futile.options_1.0.1    openxlsx_4.2.8          memoise_2.0.1          
#> INFO [2025-06-23 22:36:32] [73] bslib_0.9.0             zip_2.3.3               Rcpp_1.0.14            
#> INFO [2025-06-23 22:36:32] [76] SparseArray_1.8.0       whisker_0.4.1           xfun_0.52              
#> INFO [2025-06-23 22:36:32] [79] fs_1.6.6                pkgconfig_2.0.3